refactor: prompt-activate-history 도메인 네이밍 컨벤션 맞춤
