+++ src/services/counselings/applications/counsel-prompt-managements/counsel-prompt-managements.facade.ts
@@ -4,10 +4,10 @@ import { CounselTechniquesService } from "~counselings/domains/counsel-technique
 import { CounselTechniqueInfo } from "~counselings/domains/counsel-techniques/models/counsel-technique.info";
 import { PersonaPromptInfo } from "~counselings/domains/persona-prompts/models/persona-prompt.info";
 import { PersonaPromptsService } from "~counselings/domains/persona-prompts/persona-prompts.service";
+import { PromptActivateHistoryInfo } from "~counselings/domains/prompt-activate-history/models/prompt-activate-history.info";
+import { PromptActivateHistoryService } from "~counselings/domains/prompt-activate-history/prompt-activate-history.service";
 import { PromptVersionInfo } from "~counselings/domains/prompt-versions/models/prompt-version.info";
 import { PromptVersionsService } from "~counselings/domains/prompt-versions/prompt-versions.service";
-import { PromptActivateHistoryInfo } from "~counselings/domains/promptActivateHistory/models/promptActivateHistory.info";
-import { PromptActivateHistoryService } from "~counselings/domains/promptActivateHistory/promptActivateHistory.service";
 import { TonePromptInfo } from "~counselings/domains/tone-prompts/models/tone-prompt.info";
 import { TonePromptsService } from "~counselings/domains/tone-prompts/tone-prompts.service";
 import { AiModel } from "~proto/com/hearlers/v1/model/counsel_prompt_pb";
