+++ src/services/counselings/domains/prompt-activate-history/infrastructures/mappers/repository-prompt-activate-history-criteria.mapper.ts
@@ -1,4 +1,4 @@
-import { PromptActivateHistoryCriteriaFindMany } from "~counselings/domains/promptActivateHistory/promptActivateHistory.criteria";
+import { PromptActivateHistoryCriteriaFindMany } from "~counselings/domains/prompt-activate-history/prompt-activate-history.criteria";
 
 import { PromptActivateHistoryEntity } from "~common/system/persistences/entities/prompts/PromptActivateHistory.entity";
 import { FindManyOptions, FindOptionsOrder, FindOptionsWhere, LessThan } from "typeorm";
