+++ src/services/counselings/domains/prompt-activate-history/infrastructures/repository-prompt-activate-history.reader.ts
@@ -1,8 +1,8 @@
-import { RepositoryPromptActivateHistoryCriteriaMapper } from "~counselings/domains/promptActivateHistory/infrastructures/mappers/repository-promptActivateHistory-criteria.mapper";
-import { PromptActivateHistoryRepository } from "~counselings/domains/promptActivateHistory/infrastructures/promptActivateHistory.repository";
-import { PromptActivateHistories } from "~counselings/domains/promptActivateHistory/models/promptActivateHistory";
-import { PromptActivateHistoryCriteriaFindMany } from "~counselings/domains/promptActivateHistory/promptActivateHistory.criteria";
-import { PromptActivateHistoryReader } from "~counselings/domains/promptActivateHistory/promptActivateHistory.reader";
+import { RepositoryPromptActivateHistoryCriteriaMapper } from "~counselings/domains/prompt-activate-history/infrastructures/mappers/repository-prompt-activate-history-criteria.mapper";
+import { PromptActivateHistoryRepository } from "~counselings/domains/prompt-activate-history/infrastructures/promptActivateHistory.repository";
+import { PromptActivateHistories } from "~counselings/domains/prompt-activate-history/models/prompt-activate-history";
+import { PromptActivateHistoryCriteriaFindMany } from "~counselings/domains/prompt-activate-history/prompt-activate-history.criteria";
+import { PromptActivateHistoryReader } from "~counselings/domains/prompt-activate-history/prompt-activate-history.reader";
 
 import { Injectable } from "@nestjs/common";
 import { PromptActivateHistoryId } from "~common/shared-kernel/identifiers/prompt-activate-history.id";
