+++ src/services/counselings/domains/prompt-activate-history/prompt-activate-history.store.ts
@@ -1,12 +1,12 @@
 import {
   PromptActivateHistories,
   PromptActivateHistoriesNewProps,
-} from "~counselings/domains/promptActivateHistory/models/promptActivateHistory";
+} from "~counselings/domains/prompt-activate-history/models/prompt-activate-history";
 
 import { Injectable } from "@nestjs/common";
 
 @Injectable()
-export abstract class PromptActivateHistoryPersister {
+export abstract class PromptActivateHistoryStore {
   abstract create(promptActivateHistory: PromptActivateHistoriesNewProps): Promise<PromptActivateHistories>;
   abstract update(promptActivateHistory: PromptActivateHistories): Promise<PromptActivateHistories>;
   abstract updateMany(promptActivateHistories: PromptActivateHistories[]): Promise<PromptActivateHistories[]>;
